requirements.txt
accelerate==1.2.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
compel==2.0.3
diffusers==0.32.1
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
peft==0.14.0
pillow==11.1.0
psutil==6.1.1
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.7.0.dev20250112
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zipp==3.21.0
